Rename file, and remove repeated tests
